Extract factory members to module.
