+++ phanas_pydantic_helpers/helpers/__init__.py
@@ -1,1 +1,2 @@
+from .factory import *
 from .helpers import *
